+++ hdk-rust/wasm-test/src/lib.rs
@@ -23,7 +23,7 @@ use holochain_wasm_utils::{
     api_serialization::{
         get_entry::{GetEntryOptions, GetEntryResult},
         get_links::GetLinksResult,
-        query::QueryArgsNames,
+        query::{ QueryArgsNames, QueryArgsOptions, QueryResult },
     },
     holochain_core_types::{
         cas::content::{Address, AddressableContent},
@@ -274,8 +274,27 @@ fn handle_check_query() -> ZomeApiResult<Vec<Address>> {
     }
     let addresses = hdk::query(vec!["[%]*", "testEntryType"].into(), 0, 0).unwrap();
     if !addresses.len() == 5 {
-        return err("System Addresses not length 3");
+        return err("System + testEntryType Addresses not length 5");
     }
+
+    // Confirm same results via hdk::query_result
+    let addresses = match hdk::query_result(vec!["[%]*","testEntryType"].into(),
+                                            QueryArgsOptions::default()).unwrap() {
+        QueryResult::Addresses(av) => av,
+        _ => return err("Unexpected hdk::query_result"),
+    };
+    if !addresses.len() == 5 {
+        return err("System + testEntryType Addresses enum not length 5");
+    };
+    let headers = match hdk::query_result(vec!["[%]*","testEntryType"].into(),
+                                          QueryArgsOptions{ headers: true,
+                                                            ..Default::default()}).unwrap() {
+        QueryResult::Headers(hv) => hv,
+        _ => return err("Unexpected hdk::query_result"),
+    };
+    if !headers.len() == 5 {
+        return err("System + testEntryType Headers enum not length 5");
+    };
 
     hdk::query(QueryArgsNames::QueryName("testEntryType".to_string()), 0, 1)
 }